requirements.txt
Django
dj_database_url 
django_heroku 
djangorestframework 
django-cors-headers
django-environ
firebase_admin 
fuzzywuzzy 
pandas 
gunicorn
asyncio
